Implemented Lazy Loading
Instead of sending random movies one by one. Constructed a list and sent a whole bunch of random movies.

What is Lazy Loading?
Lazy loading is a design pattern commonly used in computer programming to defer initialization of an object until the point at which it is needed. It can contribute to efficiency in the program's operation if properly and appropriately used. The opposite of lazy loading is eager loading.